src_plugins/paella_plugin/requirements.txt
einops
rudalle
open_clip_torch
git+https://github.com/pabloppp/pytorch-tools
